adding strdup and udating putnbr_fd

+++ libft/ft_strnstr.c
@@ -6,17 +6,18 @@
 /*   By: diegrod2 <diegrod2@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/30 16:30:47 by diegrod2          #+#    #+#             */
-/*   Updated: 2024/09/30 16:30:58 by diegrod2         ###   ########.fr       */
+/*   Updated: 2024/10/15 17:45:43 by diegrod2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
+#include <stdlib.h>
 // #include <stdio.h>
 
 char *strnstr (char *big, const char *little, unsigned int len) 
 {
-	unsigned int i;
-	unsigned int j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	if (little[0] == '\0')
@@ -37,6 +38,6 @@ char *strnstr (char *big, const char *little, unsigned int len)
 // {
 // 	char *big = "Hello World!";
 // 	const char *little = "World";
-// 	unsigned int len = 12;
+// 	size_t len = 12;
 // 	printf("%s\n", strnstr(big, little, len));
 // }
